// src/store/enums/CustomerActionsEnum.ts
export enum CustomerActionsTypes {
  GET_ALL_CUSTOMERS = "GET_ALL_CUSTOMERS",
  GET_ALL_CUSTOMERS_SUCCESS = "GET_ALL_CUSTOMERS_SUCCESS",
  GET_ALL_CUSTOMERS_FAILURE = "GET_ALL_CUSTOMERS_FAILURE",

  GET_CUSTOMER_BY_ID = "GET_CUSTOMER_BY_ID",
  GET_CUSTOMER_BY_ID_SUCCESS = "GET_CUSTOMER_BY_ID_SUCCESS",
  GET_CUSTOMER_BY_ID_FAILURE = "GET_CUSTOMER_BY_ID_FAILURE",

  CREATE_CUSTOMER = "CREATE_CUSTOMER",
  CREATE_CUSTOMER_SUCCESS = "CREATE_CUSTOMER_SUCCESS",
  CREATE_CUSTOMER_FAILURE = "CREATE_CUSTOMER_FAILURE",

  UPDATE_CUSTOMER = "UPDATE_CUSTOMER",
  UPDATE_CUSTOMER_SUCCESS = "UPDATE_CUSTOMER_SUCCESS",
  UPDATE_CUSTOMER_FAILURE = "UPDATE_CUSTOMER_FAILURE",

  DELETE_CUSTOMER = "DELETE_CUSTOMER",
  DELETE_CUSTOMER_SUCCESS = "DELETE_CUSTOMER_SUCCESS",
  DELETE_CUSTOMER_FAILURE = "DELETE_CUSTOMER_FAILURE",

  GET_CUSTOMER_BY_USER_ID = "GET_CUSTOMER_BY_USER_ID",
  GET_CUSTOMER_BY_USER_ID_SUCCESS = "GET_CUSTOMER_BY_USER_ID_SUCCESS",
  GET_CUSTOMER_BY_USER_ID_FAILURE = "GET_CUSTOMER_BY_USER_ID_FAILURE",

  GET_CUSTOMERS_BY_NAME_OR_NIT = "GET_CUSTOMERS_BY_NAME_OR_NIT",
  GET_CUSTOMERS_BY_NAME_OR_NIT_SUCCESS = "GET_CUSTOMERS_BY_NAME_OR_NIT_SUCCESS",
  GET_CUSTOMERS_BY_NAME_OR_NIT_FAILURE = "GET_CUSTOMERS_BY_NAME_OR_NIT_FAILURE",
}